// asr.webapp.v2/src/app/modules/configuration/owner/owner.types.ts
export class Owner {
    ownerId: number;
    ownerCode: string;
    ownerName: string;
    ownerTypeId: number;
    ownerTypeName: string;
    status: number;
    clientIP: string;
    registerUserID: number;
    registerUserName: string;
    localChangeDate: string;
    localChangeTime: string;
    chaneDate: string;
    ownerTypeIdList: Array<number>;
}
export class OwnerType {
    ownerTypeId: number;
    ownerTypeCode: string;
    ownerTypeName: string;
    status: number;
    clientIP: string;
    registerUserID: number;
    registerUserName: string;
    localChangeDate: string;
    localChangeTime: string;
    chaneDate: string;
}